Created page that navigates
